[BPK-3068] Rename style guides do guidelines
