+++ docs/src/layouts/GuidelinesLayout/GuidelinesLayout.js
@@ -16,6 +16,20 @@
  * limitations under the License.
  */
 
-import layout from './StyleGuideLayout';
+import PropTypes from 'prop-types';
+import React from 'react';
 
-export default layout;
+import SideNavLayout from '../SideNavLayout';
+import links from '../links';
+
+const StyleGuideLayout = ({ children, ...rest }) => (
+  <SideNavLayout activeSection="GUIDELINES" links={links} {...rest}>
+    {children}
+  </SideNavLayout>
+);
+
+StyleGuideLayout.propTypes = {
+  children: PropTypes.node.isRequired,
+};
+
+export default StyleGuideLayout;
